add interface for session utility
